Move while loop to check BCLs into the BCLvalidator class Allow for run not existing in the LIMS (e.g. if it has just started), assume that it is in the sequencing stage
